Unify jump targets and instruction annotations as Relocations
There were two different notions in renovate that served similar purposes:

1. The "tag" on a `TaggedInstruction` was a symbolic target of a branch
instruction
2. The annotations on instructions captured memory addresses referred to by
load/store operations

These are both types of relocation. This change unifies the two as a single
`Relocation` type that is used as an operand annotation uniformly. This
simplified a lot of code and eliminated duplication in each backend (re-defining
the same `TargetAddress` type). The `Relocation` type is extensible by each
architecture (this is likely going to be necessary for Thumb).

Note: This commit removes the test-reporting test cases. These tests fork and
can fail in two different ways, which makes them problematic for test suite
stability.

Fixes #71