Update to the latest macaw
The macaw discovery API changed a bit in two ways:

1. There is an alternative API based on an incremental computation monad
2. The `exploreMemPointers` functions is no longer exported

The incremental API isn't quite suitable for our use here, as it doesn't expose
the partial discovery results for inspection (it only logs discovered
addresses).

We just no longer use the `exploreMemPointers` option (which corresponds to the
heuristic that treats any pointer written to memory as a potential function
pointer and discovery root). We were not relying on it anyway, so it isn't
needed.