Unify various logging and data collection APIs under one streaming logging interface
Many of the different sub-components had their own approaches to metrics
collection that were manually threaded through the entire codebase. This commit
unifies them into a single reporting mechanism via lumberjack, which enables
renovate to stream metrics and diagnostics to clients.

This commit also unifies all of the exception types into a single one. A tweak
to the definition of the exception type allows us to remove many extraneous
`Typeable` constraints. This is a major improvement because some of the ad-hoc
exception types were not exported to users, and therefore not catchable in their
structured forms.

- Convert some exceptions into panics (they were meant to indicate implementation
errors).
- Convert calls to `fail` into structured exceptions
- Push the RewriteM monad into IO so that we can stream diagnostics out (instead
of collecting them separately).  Convert instrumentation site reporting into a
stream over the logs, too, instead of collecting a list.
- Revise the nesting of the two different rewrite monads.  Rename one of the
rewrite monads to the Redirection monad. This commit untangles the two so that
rewriting (with the user-level RewriteM monad) is separate from the redirection
step.  Their connection was previously tenuous and unnecessary. Separating them
simplified both of those monads and also the call site.
- Revise the `redirect` and `runRedirectT` API to not return duplicate information
and to coalesce their returned information into a single aggregate value
- Simplify the return types of some API functions to avoid possibly missing
fields, while no longer exposing state/env types to users