Simplify the interface of the `isaJumpType` function (and related)
Instead of manually passing around the macaw `ParsedBlock`, just capture it in
ARM instructions that may need it. This simplifies all of the callers by not
requiring them to plumb the `ParsedBlock` through.

The caveats for the ARM architecture are noted (using `isaJumpType` after
rewriting can be unsafe *if* caller-inserted instructions modify PC to
jump). Note that this would have been broken before, too, but it is more
obviously the case now.