Disable the fork.c test case
Recent changes to macaw enable it to discover much more code in this binary,
including some Thumb code that renovate currently cannot
rewrite (#90). Moreover, the additional code discovered makes the test take over
an hour, so disable it for now.

We should probably replace the test with a smaller version.