Update submodules (#89)
The biggest changes were in macaw (with a few from crucible-llvm's memory model). While we are at it, update the code to build with ghc 9.0 and 9.2.

The update required disabling the fork.c test case. Recent changes to macaw enable it to discover much more code in this binary, including some Thumb code that renovate currently cannot rewrite (#90). Moreover, the additional code discovered makes the test take over an hour, so disable it for now.

We should probably replace the test with a smaller version.