Make section treated as ".text" configurable
While ".text" is commonly the name given to the Elf section containing
program code, it turns out that this isn't always the case. I
encountered a binary for PowerPC intended for the "Book E" extended
PowerPC architecture that had an empty ".text" section and instead
provided all of its program code in a section called ".text_booke".
That meant that renovate failed to find code in the binary since the
assumption that ".text" would be the right section name was hard-coded
in a couple of places. This change generalizes how the text section is
identified by making it configurable in RenovateConfig to allow for this
situation while still defaulting to ".text" to support the most common
cases.

This change:

* Adds a RenovateConfig field, rcTextSectionName, to track which section
  Renovate should treat as the ".text" section of the binary that it is
  processing.

* Updates RewriterEnv to carry the text section name as specified in
  the RenovateConfig so that the section name can be available in the
  ElfRewriter monad.

* Updates makeRewriterEnv to use the new setting when it calls
  findTextSection.

* Updates the X86_64 and PPC implementations to use ".text" in their
  default RenovateConfigs.