add isaIsRelocatableJump to Renovate.ISA
This allows an instruction to be classified as relocatable without
requiring its surrounding context. This is needed in order to classify
instructions which do not have a concrete context (i.e. they have
have a symbolic address due to some transformation, and the
original address is not readily available).

In x86_64 and PPC this is a straightforward change, since their
isaJumpType implementations only use the instruction address to
compute the relative jump target, i.e. the determination of
whether or not the instruction could be relocated was independent
of the given address.

In ARM, isaJumpType uses the given 'ParsedBlock' and instruction
address in order to use macaw's analysis of the block when determining
the jump type. To handle this, we split the ARM instruction type into
two constructors: one for no relocation annotations and one with them.
The latter has an additional flag to hold the determination of whether
or not the instruction is a relocatable jump.
When the relocation annotations are compute for an un-annotated instruction,
the flag is computed and stored in the annotated instruction constructor.